rm old vault implementation

+++ src/VaultFactory.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.20;
 
 import {IERC4626} from "openzeppelin-contracts/interfaces/IERC4626.sol";
 import {ERC20, IERC20, IERC20Metadata} from "openzeppelin-contracts/token/ERC20/ERC20.sol";
-import {Vault} from "./Vault.sol";
+import {VaultV2 as Vault} from "./VaultV2.sol";
 import {TwabController} from "pt-v5-twab-controller/TwabController.sol";
-
+// 
 contract VaultFactory {
     Vault[] public allVaults;
 
@@ -28,7 +28,7 @@ contract VaultFactory {
         IERC4626 yieldVault_,
         address claimer_,
         address yieldFeeRecipient_,
-        uint256 yieldFeePercentage_,
+        uint32 yieldFeePercentage_,
         address owner_
     ) external returns (address) {
         Vault newVault = new Vault{
